requirements.txt
pyairtable==3.0.1
python-dotenv==1.0.1
WeasyPrint==63.1
Jinja2==3.1.5
num2words==0.5.14
babel==2.16.0
requests==2.32.3
numpy==2.2.0
pillow==11.1.0
Spire.Doc==13.1.0
PySide6==6.8.1.1